requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
antiorm==1.2.1
astunparse==1.6.3
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
blinker==1.6.2
blis==0.7.10
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
chatbotAI==0.3.1.3
click==8.1.6
colorama==0.4.6
confection==0.1.1
cymem==2.0.7
daytime==0.4
distlib==0.3.7
dnspython==2.4.2
et-xmlfile==1.1.0
exceptiongroup==1.1.2
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Mail==0.9.1
flatbuffers==23.5.26
frozenlist==1.4.0
fuzzywuzzy==0.18.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
gunicorn==21.2.0
h11==0.14.0
h5py==3.9.0
idna==3.4
itsdangerous==2.1.2
jellyfish==1.0.0
Jinja2==3.1.2
joblib==1.3.1
keras==2.13.1
langcodes==3.3.0
Levenshtein==0.22.0
libclang==16.0.6
load-dotenv==0.1.0
Markdown==3.4.4
MarkupSafe==2.1.3
multidict==6.0.4
murmurhash==1.0.9
networkx==3.1
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.8
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.1
pandas==2.0.3
pathy==0.10.2
peppercorn==0.6
Pillow==10.0.0
platformdirs==3.10.0
preshed==3.0.8
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pymongo==4.4.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.22.0
pytz==2023.3
rake-nltk==1.0.6
rapidfuzz==3.3.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sampleproject==3.0.0
segtok==1.5.11
sequence==0.3.4
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.3
srsly==2.4.7
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
tflearn==0.5.0
thinc==8.1.11
tqdm==4.65.0
typer==0.9.0
tzdata==2023.3
urllib3==1.26.16
wasabi==1.1.2
websockets==11.0.3
Werkzeug==2.3.6
wrapt==1.15.0
yake==0.4.8
yarl==1.9.2
